+++ notify_send.py
@@ -46,7 +46,7 @@ SCRIPT_NAME = 'notify_send'
 SCRIPT_AUTHOR = 's3rvac'
 
 # Version of the script.
-SCRIPT_VERSION = '0.3 (dev)'
+SCRIPT_VERSION = '0.3'
 
 # License under which the script is distributed.
 SCRIPT_LICENSE = 'MIT'